Modified to also build a flat binary kernel image
It's still not working how I'd like, but I can't seem to get the
linker script to build the binary correctly without using a MEMORY
section, which I'd like to avoid.  I've also modified the boot code
so that it can start in EL2 or EL3.  Passing an elf to qemu will
start in EL3, but passing a flat binary will start in EL2, which
is what happens with the actual hardware.

I tried getting it to run on a real pi, with limited luck.  It was
working enough to print a message to the serial port, but for some
reason it's now causing an exception just after boot, and then
printing the fatal error message causes another exeception, which
loops