Modified context switching to use a struct
Previously it was using a stack to store all the registers, which
would be popped off after.  I had used the stack on computie's os,
and I still feel a bit torn on which way to go.  The advantage of
using a struct is that the translation table pointer is always
available, which would not be saved during the switch, only stored
in the system register when restoring a context.