Added a separate address space for a user process
The test process that's created is now using a separate translation
table than the kernel, with a function to map an arbitrary address
(although it only works on a single page at a time atm).

There was an issues with the stack pointer that only came up when
the user process's address space was active.  The kernel stack was
still in lower memory, and since it's not in kernel space, the cpu
will use the user space translation tables which don't have the
stack area mapped, causing a fatal error exception when contexting
switching.  The start.s code will now reset the stack pointer before
calling the rust code