Fixed the translation table mapping function
It can now map arbitrary sized segments and checks to see if the
area is already mapped, returning an error if it is.

When adding the return values, an error came up when trying to print
text after the user process syscall, which was because the address
used to print to the serial port was using user space addresses, and
I guess adding the extra code was enough to finally have the TLB
invalidated.  I need to add the instruction to manually invalidate
entries.  I also patched the address so it uses kernel space, but
I'll need to improve how I do that...