Added TLB invalidation and restoration of kernel vmem on context switch
Adding the TLB invalidation was causing a fatal exception when done
before restoring the context because the kernel context struct was
being accessed through a physical mem addr (and isn't mapped into the
user process).  It would have been possible to patch the physical
address so the kernel TLB would be used, but the alternative I went
with was to invalidate the TLB only just before the eret when restoring
a context, and also immediately copying TTBR1 to TTBR0 when entering
the kernel before saving the context.  I have to write to the kernel
stack first to save the x0 register so it can be used temporarily
which should be ok as long as the stack pointer is always using kernel
vmem.