Modified Mutex to act like a spinlock instead
In order to minimize the time spent in the critical section (with
interrupts disabled) it now uses an AtomicBool which only disables
interrupts long enough to change the value of the bool (with an
optional check before changing).  If the value is already true
(locked) then it will loop and try again.  There needs to be a bit
of delay in order to prevent excessive checking (preventing the lock
from being released by whatever else is holding it), but this is
a rare case atm, so I'll leave it for later

Also included is some code to identify the Data Abort and Instruction
Abort exception conditions which will eventually be used by the mmu,
to load the requested page.