Fixed issue with serial output on the raspberry pi
Only a few characters were being printed to the serial port when
booting off the raspberry pi, vs running it in qemu which worked
fine.  Disabling the MMU enable (and modifying the linker script to
not use the upper kernel address space) fixed the output issue, at
least until it tried to run the user process which requires MMU
address translation.  After some digging I found that the issue
was because I didn't set the MAIR_EL1 register, which controls the
caching of memory areas.  Setting this to 0 disables caching for
all virtual memory (since all the table entries use the same memory
type attribute, index 0 of MAIR_EL1).

This is not the correct solution, but I only have one 1GB block
entry in the default kernel translation tables, and this fixes
the immediate problem.  I'll have to make the kernel tables 3
levels deep instead of 2 so that I can disable caching only for
the peripheral addresses in the 0x3F00_0000 range.