Modified so only the kernel memory accesses aren't cached
I modified the value of MAIR_EL1 to include a cached memory type
with index 0, so by default the memory pointed to by the translation
tables is cached, and then I modified the kernel TTs to use the
uncached memory type (MAIR index 1, in the TT descriptor bits 4:2)
for the 1GB blocks.  This means all kernel memory usage should be
uncache and all user memory should be cached.  Again this is just
easier than making a static level 2 table for the kernel with the
entire table filled with 2MB blocks where only the 0x3F00_0000 space
uses uncached memory