Added very simple SD card reader
It's very much a minimal set of code to get the SD card reader
working in qemu only (untested on pi).  Documentation for the
Arasan SD host, included in the BCM2837 peripherals documantation
(raspberry pi 3's SoC), is pretty limited, and doesn't document the
common SD SPI command set or initialization.

A helpful resource was http://elm-chan.org/docs/mmc/mmc_e.html which
documents the commands needed, but that doesn't seem to be enough to
get data.  I had to use the read sector test from
https://github.com/bztsrc/raspi3-tutorial to verify that it could
read a sector, and then comment out code to find a minimal set of
commands.  The 3 commands issues after the configuration, which read
some values and select the card aren't documented at the first link
but seem to be necessary.