Reorganized the emmc driver
I'm still trying to figure out how I want to abstract away the
driver interface, given that I'm not sure I want to just copy
the unix style interface this time.  It still doesn't work on
the pi hardware but it gets past the SendOpCond (ACMD41) before
failing due to the command inhibit bit still being set (no amount
of timeout seems to be enough to clear it).