Converted VirtualAddress/PhysicalAddress to new types
There was a lot of unsafe pointer arithmetic and references in
various places like the page and vmalloc allocation code, without
a lot of dereferencing, so I modified it to use the Physical and
VirtualAddress types.  I converted the Address types to tuple
structs to both restrict what can be done with them (because they
still are meant to become addresses at some point), and also
provide a function to add a usize to the u64 without the need to
convert it everywhere in the code (only in the .add() method).

I also modified the mmu translation table code to use slices instead
of raw pointers, which reduces the amount of unsafe code, moving it
to the few places where slices are created.