Modified page allocator to have multiple regions
Since the available memory might not be in one contiguous chunk,
there needs to be a way to allocate pages from multiple regions.
There is now a PagePool which will check each region in turn to
find an allocated page.  There is only one region created at start
up, at the moment.

I also had an issue that came up where there was an error in
restore_context, which turned out to be because the TTRB0_EL1
register is set before restoring the context, and sometimes the
translation table cache clears itself before the TLBI instruction,
which then causes a page fault because the context struct isn't
available in user space.  To fix this, I modified the kernel heap
to use kernel space addresses, so TTBR1_EL1 will be used instead
for the lookups.  It's a bit of a hack, mostly because I'm just
hardcoding the kernel addresses when initializing the heap, but
it's cleaner than trying to restore TTBR0_EL1 only at the end.