Renamed Mutex to Spinlock
Spinlock is a more accurate name because it only disables interrupts
when locking/unlocking and uses a bool to represent the lock state.

I also added NEXT_PID to process in order to get the next pid to
use when creating a process.  It doesn't handle the case where it
overflows, but I wont be using that many processes for quite a while