Fixed emmc driver to boot on raspi3 hardware
There were a few issues.  One was that the INTERRUPT register needed
to be written to before sending a command (I assume that resets the
interrupt circuitry).  I also needed to increase the timeouts and
move the STOP_TRANSMISSON command to only be sent after all segments
of data are read.  I need to get a better way of delaying