Modified mmu mapping to use a closure like unmap
The closure is called for each granuale that is mapped and is given
the PagePool, the VirtualAddress of the granuale, and the length of
the granuale, and which optionally returns the PhysicalAddress to
map to.  If None is returned, the granuale will not be mapped and
instead decend to a smaller sized granuale.  This allows the function
to be used to make existing addresses or newly allocated 4KB pages.
I also added a MemoryType to give information for the TT descriptor
attributes, which makes it possible to map memory with the access
flag cleared so that the page is allocated on demand (although there
isn't yet a way to set the access flag for a descriptor that is
already mapped)