Reorganized to configure the IO for each platform
I tried a few methods of making a clean board-specific configuration
so that only certain drivers are compiled in, and the associated
devices can be created.  I tried having the configuration outside
the kernel, which was configured as a library, which sort of worked
but had issues with printkln macros being used both in the kernel
and the configuration.  I also tried using the bin/ directory in
the kernel/src/ but that had the same issues.

The method I've gone with for now uses the config/ directory to host
board-specific configuration files, which are selected using a single
feature flag.  The configuration then uses the path directive to
select drivers for importing, with the driver/ directory not having
a mod.rs file so that it isn't included by default.

I also added a CharDriver type and reworked the console driver a bit
but there is still more to do