Fixed boot issue caused by BSS not being zero'd
I had left out initializing the BSS segment of the executable at
boot time, but I finally had an issue caused by it.  The new console
code uses a Spinlock around the console device, which should be
initialized to false, but when booting on hardware only, it was
not printing anything because the Spinlock was locked (true) even
though it shouldn't have been.  This was just because the BSS was
not being zero'd so the compiler assumed it would be initalized to
false (0), but existing data in memory (hardware only) meant it was
actually true at the start