Modified printk to allow a device to be set
Previously it used a hardcoded module path to find the console device,
but that creates a bidirectional dependency between the kernel and the
configuration.  In an effort to eventually allow the kernel to be a
library wtih the configuration as a binary, I added a settable global
static to printk which the configuration can set.  It's a bit clumsy
but I plan to fix it later once I have the driver registration and
management code written