Modified spinlock to use core::sync::atomic
After reading up on how Aarch64 does atomic operations, I learned
that it can use cache snooping/cache exclusivity to ensure an
operation is atomic without disabling/enabling interrupts.  The
Rust core library's atomic types use those instructions without
the need for a locking mechanism that might be OS provided, so
it should be ok to use them in no_std.  I checked Redox, Sophon,
and a few other projects, and they seem to be using those atomics,
as well as Arc, and a custom crate called spinlock-rs for
synchronization (although I'm still defining my own spinlock for
now).

I only needed to remove my implementation of AtomicBool and replace
it with the core lib's version, and of course it worked fine in qemu,
but it was failing on the hardware.  After some fiddling, I got the
exception information and it was raising an implementation defined
mmu fault when executing the LDAXRB instruction.  The only reference
to the fault I could find was in the ARMv8 reference, Exclusive
access instructions and Shareable memory locations B2.9.2 which says
it could be caused by certain memory configurations which are not
cacheable.  So these atomic instructions need the cache in order to
work correctly.

I tried a whole bunch of things to get it working.  I added a level 2
translation table to the kernel space so that I could make just the
peripheral memory non-cached and the rest of kernel space cached, but
that didn't fix it.  I went through the settings for TCR_EL1 and
nothing worked.  I tried changing the MAIR_EL1 memory caching settings
and that didn't work.  I tried making the normal kernel memory
sharable with the SH attribute.  Nothing seemed to work, even though
some of the original settings weren't correct.

Finally I went through the rust-raspberrypi-os-tutorial's values for
setting up virtual memory, and they were mostly the same as what I
had, but I noticed some extra flags when setting SCTRL_EL1.  I was
just enabling the MMU bit, but it was also enable the data and
instruction caches (flags C and I), which I wasn't aware of.  Sure
enough, that was what fixed it.  I guess caching was mostly disabled
before, but I didn't notice because the kernel space wasn't set to
not cache anyways, but since these atomics need the caches in order
to guarantee no two processors are accessing the same critical memory,
it wouldn't work without the caches enabled.