Added the start of a filesystem
I'm trying to figure out how to do the driver interface, and since
char and block drivers are files, I figured I'd sort of the basics
of the virtual filesystem, with traits for the various operations.
It's mostly a copy of what I did for ComputieOS's filesystem, but
in order to take advantage of rust's reference counting, I opted
to surround vnodes with an Arc, with a dyn trait of the operations
inside.  This means I can't access common vnode data, but since
most of the common vnode data is attributes, I made a struct of
attributes which can be accessed through an ops trait method.  The
alternative of having a common vnode with a custom vnode inside,
and passing the common vnode into every ops function seemed pretty
clunky by comparison.