Added read/write to tmpfs and modified flag types
It's now possible to use open() to create a directory, and a file in
that directory, write to the file, seek to the start, and read it
back in.

I also refactored the FileFlag and FileAccess types.  Since I need
to be able to "or" the values together to create combinations that
might not have an associated enum variant, I'm now using a tuple
struct, with some consts defined in the impl block, so that they're
in the type's namespace, and can be accessed eg FileFlags::Create.
I also added some helper functions to make it easy to combine and
test flags.