Added start of devfs and partially sorted out mounting
I've added a devfs which has a root node that will allow opening of
devices from the tty subsystem based on the driver prefix and
subdevice number, and character device nodes which act as a pass
through to a specific tty subsystem device.  It's very hacky atm
but I'm just trying to get the pieces connected.

The vnodes for each device are created per-open, so there may be
multiple vnodes in memory for the same device file, but that
shouldn't affect the behaviour.

I also added a way of mounting multiple file systems in order to
mount both tmpfs and devfs, but again it's a bit hacky and incomplete.
Each vnode now has an extra trait function to get the mounted vnode
at a location.  Only the directory vnodes support it, and have a
location to store the optional node.  It's used by the lookup to
substitute if a mount node is set, but there currently isn't support
for the .. lookup, which wont accend past the mounted filesystem's
root node