Modified to use kernel space addresses in vmalloc
Previously, when context switching from a user process, it was
switching TTBR0_EL1 to the kernel translation tables so that
physical addresses can be accessed directly.  I had originally
intended to change it in order to be able to access user data
directly, but that might not be such a good idea (eg. invalid addr,
page fault, etc), but I'd still like to use kernel space addresses.

I added a special KernelVirtualAddress type which will always have
the upper bits set, and can be created from a PhysicalAddress, and
also turned into a pointer.  I then modified all the mmu code in
aarch64 to build table references using a physical address converted
to a kernel address.