Trying different ways of abstracting I/O
I've implemented a simple means of abstracting the base address for
I/O, with the system timer driver showing an example.  I tried o few
different methods in deviceio.rs which I've included for my future
reference.

I wanted something that would avoid directly using the volatile read
and write functions, and also give some type limitations.  I was
focusing on trying to type limit the register address as opposed to
limiting the values which can be written (which would also be helpful).
At first I implemented the registers for a DeviceRegisters value as
an enum that can be converted into a usize to add to the base address
but that requires a custom implementation of From<> which is a bit
clumsy, and I don't want to have to use a macro for it yet to avoid
the temptation of making the implementation complex and hard to
understand.

Another option was to have a struct of all the registers such that
the type of each field has the offset encoded in it, but this would
require having a KernelVirtualAddress for each register, which is
not ideal.

I plan to eventually change this once I know more possible solutions
without just resorting to tock-registers or a nearly-identical custom
implementation.