Added start of simple cache data structure
The idea is to stack allocate data from a Vec and once added to the
Vec, entries will either be in-used (refcount > 0) or not in use
(refcount == 0) with entries never being removed.  If an entry not
in the list is needed, an unused entry will be reused instead.

Currently it doesn't actually save the last-used order of items, so
it can't find the least-recently-used entry to recycle, but since I'm
expecting that change to be messy, I wanted to commit what I have
before adding it