Added UnownedLinkedList type, and used it in Cache
The UnownedLinkedList type is very unsafe.  Every item in the list
must be allocated before being inserted, and must live until it's
been removed from the list.  The intention is for it to be used as
a secondary list for objects that are allocated/controlled/dropped
by other means.  It's used by the Cache type, which allocates objects
from a Vec (and doesn't drop list entries although it does drop the
data stored in the entry).  I haven't tested it well, nor have I
included a lot of functionality.  I might eventually use this for
the scheduling queues if possible.