Added bufcache to block subsystem
I was finally able to get the bufcache working now that I have a
generic Cache data struct.  I will eventually make it allocate the
buffers separately so that the size can be different for different
block devices.  I currently just have a buffer-copying function for
reading which behave like the public block device read function,
but for filesystems I would ideally be able to modify the buffers
directly, which might mean a need for an Arc