Added partition table reader
I also changed the block driver offset to u64 so that on a 32-bit
platform, it would still be able to byte address a large disk. The
partition table is manually read by the EMMC driver's registration
code.  I know this would normally happen when the card is inserted
rather than on bootup only, but I don't have card detection code
yet.