Added option to get bufs from the block bufcache
Filesystems access a lot of buffers to locate file data on the disk,
so it's more efficient if the filesystem can access the buffered
data instead of copying it to yet another buffer.  It uses the
CacheArc type to refcount the bufs that are accessed.

I also added more FileAccess modes and fixed some warnings