Added byte order conversion types
I wanted to force the need to convert, so I made types for each of
the common big and little endian unsigned numbers, which are tuple
structs so that the data format is the same as their primative
types.  I modified the ext2 OnDisk structs to use the new types, and
convert when loading into the in-memory structs (since I wasn't
directly using the OnDisk versions anymore).

I checked the code produced to make sure on Aarch64 it's using the
rev instruction, despite the manual bit operations that appear in
the functions.  Writing the functions to be more reusable made the
code much more complicated, and making assembly versions of the
functions sometimes included an extra instruction compared to the
compiler-optimized generic functions.