Added lookup implementation for ext2
It's now possible to open the testapp file in /mnt/testapp on the
ext2 image.  It can't read more than 12 * 4KiB of any file or dir,
and it can't write at all, but it's somewhat functional.

In this commit, I had an issue with the device_id being incorrect,
which was due to the fact that the inode was being loaded before
the Arc was built, so the saved NonNull pointer to the moint from
the vnode was invalid.  It turns out I didn't need to save the
root node because it's cached, so I just need to call .get_inode()
in the .get_root() trait method.  If possible though, I'd like to
prevent that with perhaps a weak reference rather than a NonNull.