Fixed emmc driver on hardware raspberry pi 3
When running on raspberry pi 3 hardware, for a while, the driver has
been reading the first few sectors without issue but then timing out
during all future requests.  After messing with it for a while, I
found a few issues.  I changed the code to start a transfer so that
it performs the command request itself, and waits until the READ_READY
bit is set, rather than CMD_DONE.  I tried the ReadMultiple command
as well, but it seemed to not continue after the first block was
received (maybe I need to read/write something to continue to the
next transfer).  There was also an issue with sector vs byte offset
when reading data from the card.  In qemu, it was definetly using
byte offsets, but on the pi it was sector offsets.  After taking an
image of the sd card on the pi and passing it to qemu, qemu started
using sector offsets.  It seems to be if the card is 2GB or less, it
uses byte offsets, and if it's 4GB or more it uses sector offsets.
I'm not yet sure how to detect that, so I'm just using a 4GB image
for testing with now.  The kernel can now read the root directory and
find "testapp" in the directory tree on pi hardware