Fixed a lock contention issue with the cache
When searching the cache for an inode, it was calling the compare
function which needs to lock() the vnode it's searching in order to
get the inode number to compare with.  Since the vnode that's
currently running a vnode operation is in that list, it tries to
lock that already locked node and loops forever in the spinlock.

Firstly, I added a timeout to the spinlock which panics if it's been
a few seconds, so at least I get an error when it deadlocks.

Secondly, I modified the cache to take a key parameter which can be
used to compare with directly, using the PartialEq trait on the key,
inside the cache get function, without any locking.  For the bufcache,
the block number wasn't needed except to compare with, so it doesn't
take any more memory space, but the vnode cache will need an extra
32-bit number for the inode