Minor cleanup and bug fixes
I finally moved the global context pointer to Context in arch/
and added a Context function to switch it, hiding the unsafe code.

I moved the filesystem initialization and tests to the config file
for raspberry pi, so that filesystems can be added based on the
configuration.

I fixed a bug in cache/linkedlist where it was getting stuck in an
infinite loop because the next/prev pointers in the linkedlist where
not set to None after removing, and cache is removing/inserting each
node to update most-recently-used position.  Setting them to None
and adding a panic check that they are None before inserting fixed it.

I started moving segments to their own source file, and getting proc
more organized, and a few other things