Added writeback to generic cache and bufcache
It's now possible to write a cached entry back to disk when it's
recycled, or the cache is cleared, or a special commit() function
is called.  The cache user is responsible for tracking whether the
entry is dirty or not (in need of a writeback) (eg. bufcache handles
the dirty bit).  It's still not easy to automatically write back
entries, but at least they will eventually be written, and the
fs code that modifies bufs can, at the end, call commit on the
device to write back everything that hasn't already.

In order to give at least some automatic tracking to bufcache, I
added a special BufGuard type with Deref defined on it, which holds
an unlocked SpinlockGuard, but only allows it to be immutably
dereferenced.  Another function can be called, .lock_mut(), which
will return an unrestricted SpinlockGuard, and also mark the buf
as dirty pre-emptively.