Fixed create-image to insert the partition table
I'm using an ext2 file image with qemu for testing which replicates
the file system on the raspi3 sd card I have.  The sd card had a
basic install of raspian, with a 256MB FAT image which holds the
kernel that's loaded by the proprietary firmware, and previously
it had a 1GB ext4 partition.  That partition used fragmented block
groups, which I decided to not use to simplify the task of getting
it running, so I put an ext2 partition on instead.

So the image created by the makefile and the one on the sd card
should be the same, so I can boot on either without changing code.
Now it includes the partition table in the qemu image so that I
don't have to manually inserted, make it easier to reset the image
to a known default for write-testing