Work in Progress: ext2 writing support
It's now possible to update (but not create) the superblock,
allocate/free bits in the bitmaps for inodes and blocks, allocating
and freeing inodes themselves, and writing them back to disk, and
writing file contents.  Looking up and allocating blocks in the
block tables for a given file is broken but works for files below
12 blocks in size.  Writing directories isn't working at all.

I added a new method to the vnode operations, .commit(), which is
meant to write the vnode data back to memory.  This shouldn't be
called by anything outside of specific fs implemantations, but it's
safe to call it (it just writes the inode data to disk).  This is
the only easy way to get the exposed Ext2Vnode object from an Arc
of VnodeOperations, and is used by the cache to write dirty inodes
back to disk.

I'm not yet calling block::commit anywhere, so the data isn't
actually written to disk, but tests will print the cached bufs to
confirm that the data is correct.