Added open, close, and read syscalls
It's now possible for the testapp to open a file (written by a test
in the startup code in the kernel), read the contents of that file,
and write it to the stdout file to be printed by the kernel.

There was a bug in the file descriptor code which was overwriting
the open file entry because it was checking for a slot that was
some instead of none.

There was also a bug with the inode allocation, where the inode that
was created in the kernel test was not added to the vnode cache,
which meant it was loaded from disk when the file was opened by
testapp.  I split the cache .get() function into two, and modified
alloc_inode to .insert() the node and use the result as the return,
which fixed the issue (no data read from file when it was written
correctly)