Added file-backed on-demand page loading
The elf loader now creates segments in the processes' virtual
memory space.  When a page fault occers, the page that's allocated
can be initialized using a SegmentOperations trait which can read
its data from a file (ie. the executable which is left open).

There were issues with the page alignment, such that the testapp
binary has two segments, each separated by 64KB, but with the text
segment after the rodata segment, and with data starting at an
offset equal to the length of the rodata segment, so I had to allow
for a memory offset where the data from the file should be loaded into
the page at an offset in memory, and the first bit of memory (0x290
bytes in this case) would be unused and thus read as zero (or whatever
was written to them before).  I'm not sure if this is supposed to
be a case that needs to be handled, or if compiling the testapp is
doing wrong, or needs a custom linker script (as opposed to the
linux abi target)

The file backed loader works for the text segment but for the rodata
segment, it was not accessing the page until it passed a reference in
a syscall, so the kernel was the first thing to trigger the demand
load, but that was routed to cause a fatal error.  Trying to enable
the kernel synchronous exception to handle it resulting in only the
dedug characters being printed and not the actual messages in the
rust handler function.  This happened early on when it would print
! or $ but not the "Fatal Error..." message, but at that point, I
think it was because the fp unit was disabled, or there was an issue
with the floating point registers.  Trying to save those registers
however causes a null pointer exception, and I don't know why...