Added common irq handler with swappable controller
In order to handle and register for IRQs, I've made (still interim)
irq manager where handlers can be registered, as well as abstracting
the interrupt controller into a trait, such that it can be configured
in the config file and set during startup.  Drivers in the system can
directly access the irq manager to enable/disable their respecitive
irqs, and register a handler.

It's not doing proper bottom half processing, which I'll add later
as needed