Fixed issue with page loading
The mem_offset was causing a problem when calculating the file
offset, since it would end up being 0 instead of the correct file
offset, because the mem_offset was the same as the file offset, and
it was being subtracted from the file offset.  The proper fix was
to use saturating_sub to subtract the mem_offset from the segment
offset (which was 0), which results in the data offset being 0,
which is only then added to the file offset.  There might still be
some bugs here and I don't like this to begin with (would like
program segments in the elf executable to start on a page boundary)
but it'll do for now