Added rx support to the uart driver
It uses an IRQ to trigger a handling of the input, which is then
stored in a CircularBuffer.  Reading from the character device
interface gives you the buffered data.  It's a bit clumsy because
I've split it into two devices, a raw console device which is
static mut and which holds the buffer, and a device that's given to
the tty subsystem, which then reads the raw device.  This makes it
easier to access the buffer from the serial irq, since wrapping it
in an Arc would mean only CharOperations could be performed.

I ran into an issue when initially wrapping the raw console in a
spinlock.  That was for some reason causing things to lock up
completely when the irq occurred.  In at least some cases, it caused
a data abort, with a strange address that didn't make any sense.
It wasn't even reaching the spinlock timeout, possibly because it
was triggering other exceptions that ended in "!" being printed, but
nothing else.  This could be related to the issues with the EL1 to
EL1 context switch, which was causing me problem recently.  I'm not
sure what was causing it, so I just removed the spinlock for now