Fixed issue with kernel mode context switching
It turned out to just be too much code for the limited space of the
handler.  For the user context, the register saving and restoring
was in their own functions, but for the kernel switch it was all in
the macro, so adding separate functions fixed the issue with getting
a fatal error accessing address 0 on startup.

There is still the same issue when the elf loader doesn't pre-trigger
a page fault and allocation.  It still shows a "$" character for
the kernel switcher, but doesn't print anything after that.