Added exec() syscall, but limited
It only takes a path argument and not the command line arguments or
environment variables, but it can now (from a user program) call
exec() to load a process in place of the current one.  It just loads
itself again, as a test.  There's still a lot missing, but it's some
progress