Added fork() syscall and also println macro
It's now possible to call the fork syscall, duplicate the process,
check the return value, and continue running or call the exec
syscall depending on the result.  I've added a "sh" program that
can be run by the kernel, which then does the fork test.  It will
eventually because a simple shell.

It's just copying the segment data and duplicating the translation
tables but not the pages mapped inside them.  There is no reference
counting though, so it might actually free all the pages of the
first process address space that is destroyed, which would be a
major bug.  Currently processes aren't destroyed though

There was an issues with the translation tables which was causing
an infinite loop.  It was because the length didn't match the granuale
size, so I added some checks to make sure the length is aligned to
a page boundary, and the segment copying code does the alignment.

I also added macros for print and println to the api library, which
uses a hardcoded FileDesc to write to