Added waitpid() syscall
It's very hacky, but it works for now.  I've co-opted the status
pointer to store a dummy value which is used to know if we're calling
that syscall the first time or the second time (after a restart).
It would normally store the process's returned exit status.  It also
needs a way to actually check which process exited.  I'd like to
eventually make a generic events system for blocking and unblocking.
I also need a way to easily lookup pids