Modified process manager to use a different list
I've had an alternate implementation of a doubled linked list where
the nodes are inside an Arc<Spinlock<>> instead of being unowned.
It makes it easier to manage the scheduling and blocking queues when
the node found by the search functions can be directly inserted or
removed from the list.  And because it's all surrounded in an Arc,
the, and the Vec stores a list of Arcs, it's now possible to remove
an entry without the addresses of the nodes changing, so no corruption
and no rebuilding of the lists every time a process is removed.  This
also means I could change the Vec in process manager to a HashMap
indexed by Pid.