Modified file pointer to close on drop
I was debating whether to also add a flag to the file pointer so
that it could be forcefully closed before the references are dropped
but that would imply checking the flag before any vfs operation,
and nothing much happens when a file is closed anyways.  I'll
probably eventually just call the commit function for the block
device after a file is closed (if for writing) and that's about as
much as needs to happen at the moment.  It might be different for
device files though