Fixed a bug that was causing pages to not be freed
When searching for the Region inside the Pool of pages that a given
page belongs to, it wasn't multiplying by the page number so no
page matched and no error was raised if it wasn't found.  I'll add
a panic! to report the error later, but it requires the changes to
segments