Added refcounting to memory segments
Since processes copy the memory segments of their parent when forked,
but then free that memory before an exec, it was possible to free the
pages of a segment while they were in use by another process (before
it execs).  To avoid this, there is now reference counting of the
segments and only on the last unmapping of memory will pages of a
segment be freed.  Otherwise only the page tables (which were
duplicated when forking) will be freed.