Added ls command and a few other things
I've added the start of a coreutils crate for various command line
programs, with "ls" as the first example.  It just reads a fixed
directory, entry by entry, and prints each name out before exiting.

I also added all the KernelError variants to ApiError and added
functions to convert between them and also to and from usize so that
it can be marshalled into the registers for a syscall return.

There was a problem with page loading again, specifically it was
(I think) in an open syscall when a page fault occurred, but since
the current process was locked in order to open a file, it caused
a lock timeout when trying to change the current process's memory
space.  Adding the line in the elf loader to pre-load the first
page has fixed it for now, but maybe long term I can separate the
locking of process "files" and "space" so that they can both be
accessed/locked at the same time